backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
python-multipart==0.0.20
sqlalchemy==2.0.36
aiosqlite==0.20.0
pdfplumber==0.11.4
python-docx==1.1.2
beautifulsoup4==4.12.3
httpx==0.28.1
google-genai==1.14.0
pydantic==2.10.4
